requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
cbor2==5.8.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.1
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath==0.82.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
nodeenv==1.10.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
prisma==0.15.0
protobuf==7.34.0
pycparser==3.0
pycurl==7.45.6
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
RapidFuzz==3.14.0
referencing==0.36.2
requests==2.32.5
requestz==0.12
rich==14.3.3
rpds-py==0.27.1
scipy==1.17.1
setuptools==69.5.1
six==1.17.0
soupsieve==2.8
starlette==0.52.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.3.0
tf_keras==2.20.1
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
